nfpm.native_libs: rename package scripts->deb
Next commit will move rules into that package
