+++ 3rdparty/python/requirements.txt
@@ -33,8 +33,8 @@ node-semver==0.9.0
 
 
 # These dependencies are for scripts that rules run in an external process (and for script tests).
-aiohttp==3.12.15  # see: pants.backends.nfpm.native_libs.scripts
-aiohttp-retry==2.9.1  # see: pants.backends.nfpm.native_libs.scripts
+aiohttp==3.12.15  # see: pants.backends.nfpm.native_libs.deb
+aiohttp-retry==2.9.1  # see: pants.backends.nfpm.native_libs.deb
 elfdeps==0.2.0  # see: pants.backends.nfpm.native_libs.elfdeps
 # These dependencies are only for debugging Pants itself (in VSCode/PyCharm respectively),
 # and should never be imported.
